requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
Markdown==3.4.3
mccabe==0.7.0
mysql-connector-python==8.0.26
mysqlclient==2.2.0
protobuf==4.23.4
pycodestyle==2.10.0
pyflakes==3.0.1
pyodbc==4.0.39
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
